add confirmation box

+++ mainwindow.cpp
@@ -221,8 +221,9 @@ void MainWindow::cleanup()
 
 void MainWindow::deleteDock()
 {
-    QString selected = QFileDialog::getOpenFileName(this, tr("Select dock to delete"), QDir::homePath() + "/.fluxbox/scripts");
-    if (!selected.isEmpty()) {
+    QString selected = QFileDialog::getOpenFileName(nullptr, tr("Select dock to delete"), QDir::homePath() + "/.fluxbox/scripts");
+    if (!selected.isEmpty() && QMessageBox::question(nullptr, tr("Confirmation"),
+                                                     tr("Are you sure you want to delete %1?").arg(selected), tr("&Delete"), tr("&Cancel")) == 0) {
         QFile::remove(selected);
         cmd.run("sed -ni '\\|" + selected + "|!p' " + QDir::homePath() + "/.fluxbox/menu-mx", true);
     }
